requirements.txt
numpy==1.20.3
opencv-python==4.5.2.52
RPi.GPIO==0.7.0
